requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_openmp_mutex=4.5=3_kmp_llvm
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.11.18=pypi_0
aiosignal=1.3.2=pypi_0
attrs=25.3.0=pypi_0
blas=2.116=mkl
blas-devel=3.9.0=16_linux64_mkl
brotli-python=1.1.0=py312h2ec8cdc_2
bzip2=1.0.8=h4bc722e_7
ca-certificates=2025.1.31=hbd8a1cb_1
captum=0.8.0=pypi_0
certifi=2025.1.31=pyhd8ed1ab_0
cffi=1.17.1=py312h06ac9bb_0
charset-normalizer=3.4.1=pyhd8ed1ab_0
contourpy=1.3.2=pypi_0
cpython=3.12.10=py312hd8ed1ab_0
cuda-cudart=12.4.127=he02047a_2
cuda-cudart_linux-64=12.4.127=h85509e4_2
cuda-cupti=12.4.127=he02047a_2
cuda-libraries=12.4.1=ha770c72_1
cuda-nvrtc=12.4.127=he02047a_2
cuda-nvtx=12.4.127=he02047a_2
cuda-opencl=12.4.127=he02047a_1
cuda-runtime=12.4.1=ha804496_0
cuda-version=12.4=h3060b56_3
cycler=0.12.1=pypi_0
datasets=3.5.1=pypi_0
dill=0.3.8=pypi_0
einops=0.8.1=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.18.0=pyhd8ed1ab_0
fonttools=4.57.0=pypi_0
freetype=2.12.1=h267a509_2
frozenlist=1.6.0=pypi_0
fsspec=2025.3.0=pypi_0
giflib=5.2.2=hd590300_0
gmp=6.3.0=hac33072_2
gmpy2=2.1.5=py312h7201bc8_3
gnutls=3.6.13=h85f3911_1
h2=4.2.0=pyhd8ed1ab_0
hpack=4.1.0=pyhd8ed1ab_0
huggingface-hub=0.30.2=pypi_0
hyperframe=6.1.0=pyhd8ed1ab_0
icu=73.2=h59595ed_0
idna=3.10=pyhd8ed1ab_1
jinja2=3.1.6=pyhd8ed1ab_0
joblib=1.4.2=pypi_0
kiwisolver=1.4.8=pypi_0
lame=3.100=h166bdaf_1003
lcms2=2.16=hb7c19ff_0
ld_impl_linux-64=2.43=h712a8e2_4
lerc=4.0.0=h0aef613_1
libblas=3.9.0=16_linux64_mkl
libcblas=3.9.0=16_linux64_mkl
libcublas=12.4.5.8=he02047a_2
libcufft=11.2.1.3=he02047a_2
libcufile=1.9.1.3=he02047a_2
libcurand=10.3.5.147=he02047a_2
libcusolver=11.6.1.9=he02047a_2
libcusparse=12.3.1.170=he02047a_2
libdeflate=1.20=hd590300_0
libexpat=2.7.0=h5888daf_0
libffi=3.4.6=h2dba641_1
libgcc=14.2.0=h767d61c_2
libgcc-ng=14.2.0=h69a702a_2
libgfortran=14.2.0=h69a702a_2
libgfortran-ng=14.2.0=h69a702a_2
libgfortran5=14.2.0=hf1ad2bd_2
libhwloc=2.11.2=default_he43201b_1000
libiconv=1.18=h4ce23a2_1
libjpeg-turbo=3.1.0=hb9d3cd8_0
liblapack=3.9.0=16_linux64_mkl
liblapacke=3.9.0=16_linux64_mkl
liblzma=5.8.1=hb9d3cd8_0
liblzma-devel=5.8.1=hb9d3cd8_0
libnpp=12.2.5.30=he02047a_2
libnsl=2.0.1=hd590300_0
libnvfatbin=12.4.127=he02047a_2
libnvjitlink=12.4.127=he02047a_2
libnvjpeg=12.3.1.117=he02047a_2
libpng=1.6.43=h2797004_0
libsqlite=3.46.0=hde9e2c9_0
libstdcxx=14.2.0=h8f9b012_2
libstdcxx-ng=14.2.0=h4852527_2
libtiff=4.6.0=h1dd3fc0_3
libuuid=2.38.1=h0b41bf4_0
libwebp=1.4.0=h2c329e2_0
libwebp-base=1.4.0=hd590300_0
libxcb=1.15=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libxml2=2.12.7=hc051c1a_1
libzlib=1.2.13=h4ab18f5_6
littleutils=0.2.4=pypi_0
llvm-openmp=15.0.7=h0cdce71_0
markupsafe=3.0.2=py312h178313f_1
matplotlib=3.10.1=pypi_0
mkl=2022.1.0=h84fe81f_915
mkl-devel=2022.1.0=ha770c72_916
mkl-include=2022.1.0=h84fe81f_915
mpc=1.3.1=h24ddda3_1
mpfr=4.2.1=h90cbb55_3
mpmath=1.3.0=pyhd8ed1ab_1
multidict=6.4.3=pypi_0
multiprocess=0.70.16=pypi_0
ncurses=6.5=h2d0b736_3
nettle=3.6=he412f7d_0
networkx=3.4.2=pyh267e887_2
numpy=1.26.4=pypi_0
ocl-icd=2.3.3=hb9d3cd8_0
ogb=1.3.6=pypi_0
opencl-headers=2024.10.24=h5888daf_0
openh264=2.1.1=h780b84a_0
openjpeg=2.5.2=h488ebb8_0
openssl=3.5.0=h7b32b05_0
outdated=0.2.2=pypi_0
packaging=25.0=pypi_0
pandas=2.2.3=pypi_0
pillow=10.3.0=py312hdcec9eb_0
pip=25.0.1=pyh8b19718_0
propcache=0.3.1=pypi_0
pthread-stubs=0.4=hb9d3cd8_1002
pyarrow=20.0.0=pypi_0
pycparser=2.22=pyh29332c3_1
pyparsing=3.2.3=pypi_0
pysocks=1.7.1=pyha55dd90_7
python=3.12.3=hab00c5b_0_cpython
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.12=7_cp312
pytorch=2.5.1=py3.12_cuda12.4_cudnn9.1.0_0
pytorch-cuda=12.4=hc786d27_7
pytorch-mutex=1.0=cuda
pytorch-tabnet=4.1.0=pypi_0
pytz=2025.2=pypi_0
pyyaml=6.0.2=py312h178313f_2
readline=8.2=h8c095d6_2
regex=2024.11.6=pypi_0
requests=2.32.3=pyhd8ed1ab_1
safetensors=0.5.3=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.15.2=pypi_0
seaborn=0.13.2=pypi_0
setuptools=79.0.0=pyhff2d567_0
six=1.17.0=pypi_0
sympy=1.13.1=pypi_0
tbb=2021.13.0=hceb3a55_1
threadpoolctl=3.6.0=pypi_0
tk=8.6.13=noxft_h4845f30_101
tokenizers=0.21.1=pypi_0
torchtriton=3.1.0=py312
torchvision=0.20.1=py312_cu124
tqdm=4.67.1=pypi_0
transformers=4.51.3=pypi_0
typing_extensions=4.13.2=pyh29332c3_0
tzdata=2025.2=pypi_0
urllib3=2.4.0=pyhd8ed1ab_0
wheel=0.45.1=pyhd8ed1ab_1
xorg-libxau=1.0.12=hb9d3cd8_0
xorg-libxdmcp=1.1.5=hb9d3cd8_0
xxhash=3.5.0=pypi_0
xz=5.8.1=hbcc6ac9_0
xz-gpl-tools=5.8.1=hbcc6ac9_0
xz-tools=5.8.1=hb9d3cd8_0
yaml=0.2.5=h7f98852_2
yarl=1.20.0=pypi_0
zlib=1.2.13=h4ab18f5_6
zstandard=0.23.0=py312h66e93f0_1
zstd=1.5.6=ha6fb4c9_0
